Create the haunted house game
Based on the origin code by lecture, a main menu with "new game", "save game" and "load game" functions are added. Also some functions like show current position as "*" on the map, get a wanted inventory, go to a wanted place. Feel free to try to modify and build yourself a new game!